@@ -1,12 +1,10 @@
 name: rtsp_ffmpeg
 description: An FFMPEG based RTSP player flutter plugin. This support Almost real time RTSP playback using native code.
 version: 0.0.1
-author:
 homepage:
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
-  flutter: ">=1.10.0"
+  sdk: '>=3.2.0 <4.0.0'
 
 dependencies:
   flutter:
